Initial props passed to the secondary modification screen
