+++ src/navigation/index.js
@@ -25,7 +25,7 @@ import {AuthProvider} from './AuthProvider';
 
 const Stack = createNativeStackNavigator();
 const stackNavigatorProps = {
-  initialRouteName: 'Login',
+  initialRouteName: 'EditProvider',
   screenOptions: {
     headerShown: false,
   },
