+++ src/screens/Provider/ProviderPreview.js
@@ -2,24 +2,13 @@ import React, {useEffect, useState} from 'react';
 import {SafeAreaView} from 'react-native';
 import {Container, ContainerWhite} from '../../components/atoms';
 import {CardOrganism} from '../../components/organisms';
+import auth from '@react-native-firebase/auth';
+import userRT from '../../utils/userRT';
 
 export const ProviderPreview = ({navigation}) => {
   const [data, setData] = useState(undefined);
   const id = auth().currentUser.uid;
-  const userRT = userId => {
-    // eslint-disable-next-line react-hooks/rules-of-hooks
-    useEffect(() => {
-      const subscriber = firestore()
-        .collection('Users')
-        .doc(userId)
-        .onSnapshot(documentSnapshot => {
-          const user = documentSnapshot.data();
-          setData(user);
-        });
-      return () => subscriber();
-    }, [userId]);
-  };
-  userRT(id);
+  userRT(id, setData);
   console.log(data);
   return (
     <ContainerWhite>
